+++ src/dkplsr.jl
@@ -2,7 +2,7 @@
     dkplsr(; kwargs...)
     dkplsr(X, Y; kwargs...)
     dkplsr(X, Y, weights::Weight; kwargs...)
-    dkplsr!(X::Matrix, Y::Matrix, weights::Weight; kwargs...)
+    dkplsr!(X::Matrix, Y::Union{Matrix, BitMatrix}, weights::Weight; kwargs...)
 Direct kernel partial least squares regression (DKPLSR) (Bennett & Embrechts 2003).
 * `X` : X-data (n, p).
 * `Y` : Y-data (n, q).
@@ -101,10 +101,11 @@ function dkplsr(X, Y, weights::Weight; kwargs...)
     dkplsr!(copy(ensure_mat(X)), copy(ensure_mat(Y)), weights; kwargs...)
 end
 
-function dkplsr!(X::Matrix, Y::Matrix, weights::Weight; kwargs...)
+function dkplsr!(X::Matrix, Y::Union{Matrix, BitMatrix}, weights::Weight; kwargs...)
     par = recovkw(ParKplsr, kwargs).par
     @assert in([:krbf ; :kpol])(par.kern) "Wrong value for argument 'kern'." 
     Q = eltype(X)
+    isa(Y, BitMatrix) ? Y = convert.(Q, Y) : nothing
     p = nco(X)
     q = nco(Y)
     xscales = ones(Q, p)