+++ src/vip.jl
@@ -37,12 +37,10 @@ Editions Technip, Paris.
 ```julia
 using Jchemo
 
-X = [1. 2 3 4 ; 4 1 6 7; 12 5 6 13; 
-    27 18 7 6 ; 12 11 28 7] 
-Y = [10. 11 13 ; 120 131 27 ; 8 12 4; 
-    1 200 8 ; 100 10 89] 
+X = [1. 2 3 4; 4 1 6 7; 12 5 6 13; 27 18 7 6 ; 12 11 28 7] 
+Y = [10. 11 13; 120 131 27; 8 12 4; 1 200 8; 100 10 89] 
 y = Y[:, 1] 
-ycla = [1 ; 1 ; 1 ; 2 ; 2]
+ycla = [1; 1; 1; 2; 2]
 
 nlv = 3
 model = plskern(; nlv)
@@ -119,5 +117,3 @@ function vip(object::Union{Pcr, Plsr}, Y; nlv = nothing)
     imp = sqrt.(A / B)
     (imp = imp, W2, rdd)
 end
-
-